Move woocommerce/product-collapsible to blocks folder
