+++ packages/js/product-editor/src/blocks/collapsible/block.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
-	"name": "woocommerce/collapsible",
+	"name": "woocommerce/product-collapsible",
 	"title": "Collapsible",
 	"category": "widgets",
 	"description": "Container with collapsible inner blocks.",
