+++ packages/js/product-editor/src/blocks/collapsible/edit.tsx
