+++ packages/js/product-editor/src/blocks/collapsible/index.ts
